+++ zhalilov.rail/P4/countNonZeroDiags.hpp
@@ -1,9 +1,11 @@
 #ifndef COUNT_NONZERO_DIAGS_HPP
 #define COUNT_NONZERO_DIAGS_HPP
 
+#include <cstddef>
+
 namespace zhalilov
 {
-  int countNonZeroDiags(const int matrix[], int rows, int cols);
+  size_t countNonZeroDiags(const int matrix[], int rows, int cols);
 }
 
 #endif
